Give `PlaybackControls` a `duration` prop
